Code cleanup and changed from HashMap to TreeMap in command manager. Made give command not require a mention as an argument.

+++ HashBot/src/main/java/me/afarrukh/hashbot/entities/Invoker.java
@@ -98,7 +98,7 @@ public class Invoker {
 
     public int getPercentageExp() {
         int exp = (int) getExp();
-        int expToProgress = (int) getExpForNextLevel();
+        int expToProgress = getExpForNextLevel();
         return (int) Math.round((double) exp/expToProgress*100);
     }
